pyproject.toml
[project]
name = "mats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "flask>=3.0.0",
    "pandas>=2.0.0",
    "transformers>=4.57.3",
    "torch>=2.9.1",
    "sae-lens>=6.24",
    "typeguard>=4.4.4",
]
